improv: parallelize some iterators, fixes: ECS pattern improvements

+++ src/strokes/chrono_comp.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, Eq, Ord, PartialEq, PartialOrd)]
 pub struct ChronoComponent {
-    pub t: u64,
+    pub t: u32,
 }
 
 impl Default for ChronoComponent {
@@ -12,7 +12,7 @@ impl Default for ChronoComponent {
 }
 
 impl ChronoComponent {
-    pub fn new(t: u64) -> Self {
+    pub fn new(t: u32) -> Self {
         Self {
             t,
         }
